ver. 2

+++ src/repository/contacts.py
@@ -1,56 +1,60 @@
+from sqlalchemy import select, func, or_, extract
+from sqlalchemy.ext.asyncio import AsyncSession
+
+from src.entity.models import Contact, User
 from typing import List
 
-from sqlalchemy.orm import Session
-from sqlalchemy import select, func, or_
+from datetime import date, timedelta
 
-from src.database.models import Contact
-from src.schemas import ContactModel
-from sqlalchemy import select
+#from sqlalchemy.orm import Session
+from src.schemas.contacts import ContactModel
 
 
-async def get_contacts(skip: int, limit: int, db: Session) -> List[Contact]:
-    result =  db.query(Contact).offset(skip).limit(limit).all()
-    print(f"))))))))))))))))))))))))))))))))))))))){result}")
 
-    db.query(Contact).offset(skip).limit(limit).all()
-    return result
+async def get_contacts(limit: int, offset: int, db: AsyncSession, user: User) -> List[Contact]:
+    stmt = select(Contact).filter_by(user=user).offset(offset).limit(limit)
+    contacts =  await db.execute(stmt)
+    return contacts.scalars().all()
    
 
 
-async def get_contact(contact_id: int, db: Session) -> Contact:
-    return db.query(Contact).filter(Contact.id == contact_id).first()
+async def get_contact(contact_id: int, db: AsyncSession, user: User) -> Contact:
+    stmt = select(Contact).filter_by(user_id=user.id).filter_by(id=contact_id)
+    contact =  await db.execute(stmt)
+    return contact.scalar_one_or_none()
 
 
-async def get_contact_firstname(firstname:str  ,  db: Session):
-    print(f"=========================================================={firstname}")
-    result = db.query(Contact).filter(Contact.firstname == firstname).all()
-    print(f"((((((((((((((((((((((((((((((((((((((((({result}")
-    return result
-    #return db.query(Contact).filter(Contact.firstname.like("%1%"))
-    # if firstname:
-    #     field = Contact.firstname
-    #     value = firstname
-    # elif lastname:
-    #     field = Contact.lastname
-    #     value =  lastname
-    # elif email:
-    #     field = Contact. email
-    #     value =  email
-    #return db.query(Contact).filter(Contact.firstname == firstname).count()
-    
-     
+async def get_contact_firstname(limit: int, offset: int, firstname: str ,lastname: str,email: str  ,  db: AsyncSession, user: User):
+    stmt = select(Contact).filter_by(user=user).where(or_(Contact.firstname == firstname,Contact.lastname == lastname,Contact.email == email)).offset(offset).limit(limit)
+    contacts =  await db.execute(stmt)
+    return contacts.scalars().all()  
 
+async def get_contact_birthday(skip: int, limit: int,  db: AsyncSession, user: User):
+    date_now = date.today()
+    year_now = date_now.year
+    diff = date_now + timedelta(days = 7)
+    date_to = date(year=year_now, month=diff.month, day=diff.day)
+    stmt = select(Contact).filter_by(user=user).\
+        filter((extract('month', Contact.databirthday) >= date_now.month)).\
+        filter((extract('month', Contact.databirthday) <= date_to.month)).\
+        filter(extract('day', Contact.databirthday) >= date_now.day).\
+        filter(extract('day', Contact.databirthday) <= date_to.day).\
+        offset(skip).limit(limit)
+    result =  await db.execute(stmt)
+    return result.scalars().all()  
 
-async def create_contact(body: ContactModel, db: Session) -> Contact:
-    contact = Contact(firstname=body.firstname, lastname=body.lastname, email=body.email, mobilenamber=body.mobilenamber, databirthday=body.databirthday, note=body.note )
+async def create_contact(body: ContactModel, db: AsyncSession, user: User):
+    contact = Contact(**body.model_dump(exclude_unset=True), user=user )
     db.add(contact)
-    db.commit()
-    db.refresh(contact)
+    await db.commit()
+    await db.refresh(contact)
     return contact
 
 
-async def update_contact(contact_id: int, body: ContactModel, db: Session) -> Contact :
-    contact = db.query(Contact).filter(Contact.id == contact_id).first()
+async def update_contact(contact_id: int, body: ContactModel, db: AsyncSession, user: User) -> Contact :
+    stmt = select(Contact).filter_by(user_id=user.id).filter_by(id=contact_id)
+    result = await db.execute(stmt)
+    contact = result.scalar_one_or_none()
     if contact:
         contact.firstname = body.firstname
         contact.lastname = body.lastname
@@ -58,17 +62,17 @@ async def update_contact(contact_id: int, body: ContactModel, db: Session) -> Co
         contact.mobilenamber = body.mobilenamber
         contact.databirthday = body.databirthday
         contact.note = body.note
-        db.commit()
+        await db.commit()
+        await db.refresh(contact)
     return contact
 
 
-async def remove_contact(contact_id: int, db: Session)  -> Contact :
-    contact = db.query(Contact).filter(Contact.id == contact_id).first()
-    db.delete(contact)
-    db.commit()
+async def remove_contact(contact_id: int, db: AsyncSession, user: User)  -> Contact :
+    stmt = select(Contact).filter_by(user=user).filter_by(id=contact_id)
+    contact = await db.execute(stmt)
+    contact = contact.scalar_one_or_none()
+    if contact:
+        await db.delete(contact)
+        await db.commit()
     return contact
 
-
-
-#)))))))))))))))))))))))))))))))))))))))[<src.database.models.Contact object at 0x0000017343CEDE90>, <src.database.models.Contact object at 0x0000017341D483D0>, <src.database.models.Contact object at 0x0000017343CEDF50>]
-#((((((((((((((((((((((((((((((((((((((([<src.database.models.Contact object at 0x000001F92DA777D0>, <src.database.models.Contact object at 0x000001F92DA77850>]
